requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
anthropic==0.28.0
pdfplumber==0.11.0
PyMuPDF==1.24.3
tiktoken==0.7.0
tavily-python==0.3.3
python-docx==1.1.2
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
sse-starlette==2.1.0
